+++ build-mac.sh
@@ -1,0 +1,7 @@
+export GEODE_BINDINGS_REPO_PATH=~/bindingsneverchange
+#cmake -B buildrwdi -DCMAKE_BUILD_TYPE=RelWithDebInfo
+#cmake --build buildrwdi --config RelWithDebInfo
+#cmake -B buildrel -DCMAKE_BUILD_TYPE=Release
+#cmake --build buildrel --config Release
+#cmake -B builddeb -DCMAKE_BUILD_TYPE=Debug
+#cmake --build builddeb --config Debug