+++ src/hacks.hpp
@@ -416,10 +416,10 @@ class Hacks {
         }
     }
     static void setPitch(float pitch) {
+#ifndef GEODE_IS_MACOS // TODO: figure out why it turns to creepypasta when tuning down speedhack
         if (!Hacks::isHackEnabled("Speedhack Audio")) {
             pitch = 1.0F;
         }
-#ifndef GEODE_IS_MACOS // TODO: figure out why it turns to creepypasta when tuning down speedhack
         auto fmod = FMODAudioEngine::sharedEngine();
         FMOD_RESULT result; // ensuring
         FMOD::ChannelGroup* master_group = nullptr;